Orders from a Coffee Shop
►Description:

The dataset used belongs to a Coffee shop that has different branches in United States, Ireland and United Kingdom between 2019 and 2022. This provides different insights from the customer behaviour regarding their products. The database (Original Data Coffee Orders.xlsx) contained the following key features distributed in different sheets.

► Key features:

• Order ID
• Order Date
• Customer ID
• Product ID
• Quantity
• Customer Name
• Email
• Address, City, Country, Postcode
• Loyalty card

► Questions:

A) Which was the total sales of each country by customers with and without loyalty card between 2020 and 2021?
The total sales by country from clients between 2021 and 2022 with loyalty card was $9,114 for US, $1,872 and $552 for United Kingdom. 
Unlikely the clients without Loyalty card, were the total sales is a bit bigger with $11,344 for United States, $1,835 for Ireland and $1,165 for United Kingdom.
A way to have more sells for clients with Loyalty card could be providing more benefits that incentivates the buying of coffee.

B) Which was the month with more sales of 1kg Liberica coffee?
It was on November 2021.

C) Who were the top 3 clients between April to June on 2022?
The top 3 clients between April and June of 2022 are Teddi Orowthe with $205 of buying products. After, the second place was Faunie Brigham with a total of $164 on buys and, finally, Doll Beauchamp with $164.

► Log from the Project Coffee Orders Data:

- Original data is shown at "Original Data Coffe Orders.xlsx"
- Data Cleaning and normalization from different sheets. Date, number and currency formatting. Check and erasing duplicate data.
- At [Orders] sheet, were added columns "Coffe Type Name", "Roast Type Name" and "Loyalty Card" through the use of "Xlookup", "If" & dynamic formulas like "Index" + "Match" from [Customers] sheet.
- Creation of a main dashboard that shows a Pivot Tables & Slicers.
- Slicers where created to filter through the order date, the roast type, the size and the loyalty card.
- Pivot charts shows the evolution of the sales by type of coffee from 2019 to 2022 by a Line Chart. 
- The second Pivot chart contains the total sales by country in USD from United States, Ireland and United Kingdom demonstrated by a Clusted bar chart
- The last Pivot chart contains the list of the top 5 clients from the shop.